🔨 extract DiscreteBars component
